requirements.txt
# host/ Python dependencies
numpy>=1.24
pyserial>=3.5
